refactor(pccp): flattenContent 공유 유틸 추출 (PR #221 리뷰 DRY)
pdf.tsx와 docx.ts에 동일한 flattenContent 로직이 중복 구현되어 있었음.
lib/pccp/exporters/content-flatten.ts로 추출, 양쪽에서 import.

MoAI <email@mo.ai.kr>
